# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Freerock Solar & Energy API",
    "version": "1.0.0",
    "description": "Production PWA for Zimbabwe's trusted solar partner — instant quotes, Paynow payments, training enrollment, admin dashboard.\n\nBase URL: `https://freerock.co.zw` (production) or `http://localhost:3000` (development).\n\nAll endpoints that accept JSON use `Content-Type: application/json`. Protected admin routes require a valid Supabase session cookie (set via `/api/auth/otp` or admin login)."
  },
  "servers": [
    { "url": "https://freerock.co.zw", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
...
  "paths": {
...
    "/api/custom-design": {
      "post": {
        "tags": ["Custom Design"],
        "summary": "Submit a custom solar design request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "phone", "propertySize", "energyUsage", "budget"],
                "properties": {
                  "name": { "type": "string" },
                  "phone": { "type": "string" },
                  "propertySize": { "type": "string", "description": "Square metres" },
                  "energyUsage": { "type": "string", "description": "Monthly kWh" },
                  "budget": { "type": "number", "minimum": 100 },
                  "need3D": { "type": "boolean" },
                  "needFinancialAnalysis": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Design request created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "request": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "customer": { "$ref": "#/components/schemas/Customer" },
                        "design": { "type": "object" },
                        "status": { "type": "string" },
                        "createdAt": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation failed" },
          "429": { "description": "Rate limited" }
        }
      }
    },
    "/api/audit-repair": {
      "post": {
        "tags": ["Audit & Repair"],
        "summary": "Book an energy audit or repair",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "phone", "serviceId", "date", "location"],
                "properties": {
                  "name": { "type": "string" },
                  "phone": { "type": "string" },
                  "serviceId": { "type": "string", "enum": ["energy-audit", "financial-audit", "system-diagnosis", "fault-repair"] },
                  "date": { "type": "string", "format": "date" },
                  "location": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Booking created" },
          "400": { "description": "Validation failed" },
          "429": { "description": "Rate limited" }
        }
      }
    },
    "/api/payment/initiate": {
      "post": {
        "tags": ["Payments"],
        "summary": "Initiate a Paynow payment",
        "description": "Creates a Paynow checkout link or returns a simulation response when Paynow credentials are absent (dev mode).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["quoteId", "paymentMethod", "amount", "customer"],
                "properties": {
                  "quoteId": { "type": "string" },
                  "paymentMethod": { "type": "string", "enum": ["ecocash", "innbucks", "telecash", "bank_transfer", "card"] },
                  "amount": { "type": "number", "exclusiveMinimum": 0 },
                  "customer": {
                    "type": "object",
                    "required": ["name", "phone"],
                    "properties": {
                      "name": { "type": "string" },
                      "phone": { "type": "string" },
                      "email": { "type": "string", "format": "email" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment initiated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payment_url": { "type": "string", "nullable": true, "description": "Paynow checkout URL. Null when in simulation mode." },
                    "simulated": { "type": "boolean" },
                    "reference": { "type": "string" },
                    "amount": { "type": "number" },
                    "paymentMethod": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation failed" },
          "502": { "description": "Paynow gateway error" }
        }
      }
    },
...
    "/api/admin/leads": {
      "get": {
        "tags": ["Admin"],
        "summary": "List all quotes as leads (admin)",
        "description": "Returns normalized lead records flat enough for table rendering. Supports filtering by status, date range.",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "leads": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Lead" }
                    },
                    "total": { "type": "integer" },
                    "source": { "type": "string", "enum": ["supabase", "mock"] }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Customer": {
        "type": "object",
        "properties": {
          "fullName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string" },
          "province": { "type": "string" },
          "propertyType": { "type": "string" },
          "roofType": { "type": "string" }
        }
      },
...
      "Lead": {
        "type": "object",
        "properties": {
          "quoteId": { "type": "string" },
          "name": { "type": "string" },
          "phone": { "type": "string" },
          "email": { "type": "string" },
          "service": { "type": "string" },
          "tier": { "type": "string" },
          "location": { "type": "string" },
          "depositStatus": { "type": "string", "enum": ["deposit_paid", "pending", "cancelled", "completed"] },
          "date": { "type": "string", "format": "date" },
          "totalUsd": { "type": "number" },
          "depositUsd": { "type": "number" }
        }
      }
    }
  }
}